+++ src/main/java/it/hysen/springmvc/dao/ArticleDAO.java
@@ -2,6 +2,6 @@ package it.hysen.springmvc.dao;
 
 import it.hysen.springmvc.model.Article;
 
-public interface ArticleDAO extends GenericDAO<Article, Long>, SoftOperationsDAO<Article, Long> {
-	
+public interface ArticleDAO extends GenericDAO<Article, Integer>, SoftOperationsDAO<Article, Integer> {
+
 }
