+++ src/main/java/it/hysen/springmvc/dao/UserDAO.java
@@ -2,14 +2,14 @@ package it.hysen.springmvc.dao;
 
 import it.hysen.springmvc.model.User;
 
-public interface UserDAO extends GenericDAO<User, Long>, SoftOperationsDAO<User, Long> {
-
+public interface UserDAO extends GenericDAO<User, Integer>, SoftOperationsDAO<User, Integer> {
+	
 	User findByEmail(User entity);
-
+	
 	User findByUsername(User entity);
-
+	
 	User findByUsernameAndPassword(User entity);
-	
+
 	String hashPassword(String password);
-	
+
 }
