+++ src/main/java/it/hysen/springmvc/service/ArticleService.java
@@ -2,8 +2,8 @@ package it.hysen.springmvc.service;
 
 import it.hysen.springmvc.model.Article;
 
-public interface ArticleService extends GenericService<Article, Long>, SoftOperationsService<Article, Long> {
-
+public interface ArticleService extends GenericService<Article, Integer>, SoftOperationsService<Article, Integer> {
+	
 	boolean isArticleExist(Article entity);
-
+	
 }
