+++ src/main/java/it/hysen/springmvc/service/UserService.java
@@ -2,12 +2,12 @@ package it.hysen.springmvc.service;
 
 import it.hysen.springmvc.model.User;
 
-public interface UserService extends GenericService<User, Long>, SoftOperationsService<User, Long> {
-
+public interface UserService extends GenericService<User, Integer>, SoftOperationsService<User, Integer> {
+	
 	User authenticate(User entity);
-	
+
 	String hashPassword(String password);
-	
+
 	boolean isUserExist(User entity);
-	
+
 }
